+++ test/mixed.js
@@ -974,7 +974,7 @@ describe('Mixed Types ', () => {
       await inst.isValid({ prop: null }).should.eventually().equal(true);
     });
 
-    it('should pass when value is not undefined nor null', async () => {
+    it('should pass when value is not undefined', async () => {
       let inst = object({
         prop: string().defined(),
       });